requirements.txt
# Core HTTP and async libraries
httpx==0.27.0
aiohttp==3.9.1
requests==2.31.0

# Database
aiosqlite==0.19.0

# Crypto libraries for API authentication
cryptography==42.0.0
pycryptodome==3.20.0

# OpenRouter/OpenAI client
openai==1.51.2
anthropic

# Data handling and analysis - updated for Python 3.13 compatibility
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.12.0       # For portfolio optimization (Python 3.13 compatible)

# Environment variables
python-dotenv==1.0.0

# Pydantic and data validation
# Using a version compatible with Python 3.13
pydantic==2.8.2
pydantic-settings

# Logging and monitoring
structlog==23.2.0

# Type hints - compatible with openai requirements
typing-extensions>=4.11

# Date/time handling
python-dateutil==2.8.2
pytz==2023.3

# Rate limiting
ratelimit==2.2.1

# Configuration management
pyyaml==6.0.1

# Error tracking and monitoring
sentry-sdk==1.39.2

# xAI Grok API client
xai_sdk>=1.0.0

# Polymarket CLOB client
py-clob-client>=0.34.5

# WebSocket streaming
websockets>=12.0

# News & sentiment pipeline
feedparser>=6.0.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.12.1
isort==5.13.2
uvicorn
json-repair